requirements.txt
tweepy==3.10.0
requests
python-dotenv